% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSquares.R
\name{addSquares}
\alias{addSquares}
\title{Sum of squares}
\usage{
addSquares(x, y)
}
\arguments{
\item{x}{a \code{numeric} vector in size n}

\item{y}{a \code{numeric} vector in size n}
}
\value{
a vector of type \code{numeric} in size n
}
\description{
The \code{addSquares} function takes the sum of squares of two
variables
}
\examples{
# call data
data(iris)

# define the function paramaters
x <- iris$Sepal.Length
y <- iris$Petal.Length

# execute the function
addSquares(x = x, y = y)

}
\author{
Necla Kochan
}
