% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substractSquares.R
\name{substractSquares}
\alias{substractSquares}
\title{Difference of squares}
\usage{
substractSquares(x, y)
}
\arguments{
\item{x}{a \code{numeric} vector in size n}

\item{y}{a \code{numeric} vector in size n}
}
\value{
a vector of type \code{numeric} in size n
}
\description{
The \code{addSquares} function takes the difference of squares
of two variables
}
\examples{
# call data
data(iris)

# define the function paramaters
x <- iris$Sepal.Length
y <- iris$Petal.Length

# execute the function
substractSquares(x = x, y = y)

}
\author{
Necla Kochan
}
